vision_services/requirements.txt
geoai-py
transformers
geopandas
rasterio
opencv-python
fastapi
websockets
